+++ data/Release.lua
@@ -4,7 +4,7 @@
 Nexus = Nexus or {}
 
 Nexus.Release = {
-    version = "1.96.1",
+    version = "1.96.2",
     baseVersion = "1.19.5",
     published = true,
     emergencyCommunityOff = false,